+++ spec/unit/resource/api1000/c7000/logical_interconnect_group_spec.rb
@@ -1,0 +1,16 @@
+require 'spec_helper'
+
+RSpec.describe OneviewSDK::API1000::C7000::LogicalInterconnectGroup do
+  include_context 'shared context'
+
+  it 'inherits from OneviewSDK::API800::C7000::LogicalInterconnectGroup' do
+    expect(described_class).to be < OneviewSDK::API800::C7000::LogicalInterconnectGroup
+  end
+
+  describe '#initialize' do
+    it 'should be initialize the instance with default values' do
+      item = described_class.new(@client_1000)
+      expect(item['type']).to eq('logical-interconnect-groupV6')
+    end
+  end
+end